fix: rerender overlay on draggable prop change for correct drang implementation
Current problem:
We have map, where we can draw different figures and place different markers. Each mode can be enabled and disabled. 
Markers can be dragged  in the specific mode.
At the moment if we change draggable props, which is passed to the marker behaviour is incorrect and deep comparison for rerendering overlay is done only with lar and lng props.